Adds new linux-specific unit test in a new file
